testing/mochitest/baselinecoverage/chrome/test_baselinecoverage.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->

<?xml-stylesheet href="chrome://global/skin/" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script type="application/javascript"><![CDATA[
    SimpleTest.ok(true, "Collecting baseline coverage for XUL (.xul) file types.");
  ]]></script>

  <body xmlns="http://www.w3.org/1999/xhtml" style="height: 300px; overflow: auto;" />
</window>
